got priority and progress to work

+++ app/src/main/java/com/example/fa11en/watoplan/data/datatype/UserEvent.kt
@@ -56,7 +56,7 @@ class UserEvent (@ColumnInfo(name = "typename") val typeName: String) {
             ParameterTypes.LOCATION -> value is Location
             ParameterTypes.ENTITIES -> value is MutableList<*>
             ParameterTypes.REPEAT -> value is MutableList<*>
-            ParameterTypes.PROGRESS -> value is Double && value >= 0 && value <= 100
+            ParameterTypes.PROGRESS -> value is Int && value >= 0 && value <= 100
             ParameterTypes.PRIORITY -> value is Int
         }
     }
